fix: User에 referenceGroupId필드 Integer -> Long으로 타입 변경

+++ src/main/java/com/map/gaja/user/domain/model/User.java
@@ -32,7 +32,7 @@ public class User extends BaseTimeEntity {
     @Column(nullable = false)
     private LocalDateTime lastLoginDate;
 
-    private Integer referenceGroupId;
+    private Long referenceGroupId;
 
     public void checkCreateGroupPermission() {
         if (authority.getGroupLimitCount() <= groupCount) {
